Move store hooks and exporters into different folder structures
