+++ src/common/utils/custom_templates.js
@@ -1,7 +1,7 @@
 import { cloneDeep } from 'lodash'
 import { remote } from 'electron'
 import { t } from 'plottr_locales'
-import { customTemplatesStore } from './store_hooks'
+import { customTemplatesStore } from '../../file-system/stores'
 import { tree, helpers } from 'pltr/v2'
 import { saveCustomTemplate } from '../../dashboard/utils/templates_from_firestore'
 const { app } = remote
