+++ src/common/utils/error_reporter.js
@@ -4,7 +4,7 @@ import path from 'path'
 import log from 'electron-log'
 import { ActionTypes } from 'pltr/v2'
 import { t as i18n } from 'plottr_locales'
-import USER from './user_info'
+import { USER } from '../../file-system/stores'
 const app = remote.app
 
 let previousAction = null
