+++ src/common/utils/full_error_report.js
@@ -3,9 +3,7 @@ import fs from 'fs'
 import path from 'path'
 import { machineIdSync } from 'node-machine-id'
 import { t } from 'plottr_locales'
-import USER from './user_info'
-import { trialStore } from './store_hooks'
-import SETTINGS from './settings'
+import { SETTINGS, USER, trialStore } from '../../file-system/stores'
 import log from 'electron-log'
 import { is } from 'electron-util'
 const { app, dialog } = remote
