+++ src/common/utils/rollbar.js
@@ -1,7 +1,7 @@
 const Rollbar = require('rollbar')
 import { remote } from 'electron'
 const app = remote.app
-import USER from './user_info'
+import { USER } from '../../file-system/stores'
 
 export default function setupRollbar(where) {
   let environment = process.env.NODE_ENV == 'development' ? 'development' : 'production'
