--- src/common/utils/user_info.js
@@ -1,4 +1,0 @@
-import Store from 'electron-store'
-import { USER_INFO_PATH } from './config_paths'
-const USER = new Store({ name: USER_INFO_PATH })
-export default USER
