+++ src/exporter/scrivener/v2/bare_scrivx.json
