+++ src/exporter/scrivener/v2/exporters/characters.js
