+++ src/exporter/word/exporters/characters.js
