+++ src/exporter/word/exporters/customAttributes.js
