+++ src/exporter/word/exporters/itemAttachments.js
