+++ src/exporter/word/exporters/places.js
